NPDOTF-54: Increase TLS buffer size in mbedtls configuration to 4608.
To save RAM, we configure the maximum TLS record size smaller than the
standard TLS default.

We can control the maximum size of a TLS record for everything except the
TLS handshake (mbedtls doesn't support fragmentation of handshake messages).
Currently, the largest handshake message needs ~4096 bytes for the certs we are
using, but we need to allow plenty of slack against possible cert changes.