NPDOTF-58: Fix STM32F4 toolchain setup to run w/o access to our internal mirror.
Add support for setting up the STM32F4 buld environment without access to
the internal Verizon mirror.  Specifying a username parameter retains the old
behavior and assumes you have SSH access to the internal mirror as that user.

Otherwise, the toolchain file is downloaded from a public repository.
The .zip file for the STM32CubeF4 library must also be manually
pre-downloaded since you have to be logged into the STMicro support site to
access it.

Moved the openocd config files that were previously downloaded from the mirror
so they now reside in this repo.

Changed common.mk to use whatever version of STM32CubeF4 is in the tools
directory, rather than being hardwired.